+++ spec/rails_app/app/components/tree/node_state.rb
@@ -1,0 +1,4 @@
+module Tree
+  class NodeState < Crud
+  end
+end