+++ spec/rails_app/db/schema.rb
@@ -47,6 +47,7 @@ ActiveRecord::Schema.define(version: 20150414045121) do
 
   create_table "file_records", force: :cascade do |t|
     t.boolean  "is_dir",     default: false
+    t.boolean  "expanded",   default: false
     t.string   "name",                       null: false
     t.integer  "size",       default: 0
     t.integer  "parent_id"