+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.12.7 - WIP
-# Fix the strong_default_attrs option support
-# Support column action's isDisabled option
+* Fix the strong_default_attrs option support
+* Support column action's isDisabled option
+* Store tree nodes expand/collapse state
 
 # 0.12.6 - 2015-07-11
 * Regression: show cell editor on adding/editing a record in grid